requirements.txt

langchain
langchain-core
langchain-community
langchain-experimental
langchain-groq
langchain-text-splitters
langchain-tavily
langchain-huggingface
langchainhub
langgraph

groq>=0.4.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
numpy>=1.26.4
pydantic>=2.5.3
streamlit>=1.32.2
streamlit-chat>=0.1.1
pandas>=2.2.0
matplotlib>=3.8.3
tavily-python>=0.2.8
sentence-transformers>=2.2.2
huggingface-hub>=0.20.3
transformers>=4.35.0
torch>=2.0.0 